requirements.txt
# Core
protobuf>=4.25.0
grpcio-tools>=1.60.0

# Kafka
confluent-kafka>=2.3.0

# Media / ML
torch>=2.1.0
faster-whisper>=0.10.0
pyannote.audio>=3.1.1
pyannote.core>=5.0.0

# Utilities
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0